+++ src/Service/Kms/src/KmsClient.php
@@ -473,6 +473,13 @@ class KmsClient extends AbstractApi
                     'signService' => 'kms',
                     'signVersions' => ['v4'],
                 ];
+            case 'me-central-1-fips':
+                return [
+                    'endpoint' => 'https://kms-fips.me-central-1.amazonaws.com',
+                    'signRegion' => 'me-central-1',
+                    'signService' => 'kms',
+                    'signVersions' => ['v4'],
+                ];
             case 'me-south-1-fips':
                 return [
                     'endpoint' => 'https://kms-fips.me-south-1.amazonaws.com',